refactor: split native provider wrapper
