+++ packages/react-native/src/provider/GTProvider.tsx
@@ -1,22 +1,10 @@
-import {
-  I18nStore,
-  type InternalGTProviderProps,
-  InternalGTProvider,
-  getReactI18nCache,
-} from '@generaltranslation/react-core/context';
-import { Suspense, use, useCallback, useMemo, useRef, useState } from 'react';
+import type { InternalGTProviderProps } from '@generaltranslation/react-core/context';
+import { Suspense } from 'react';
 import type { ReactNode } from 'react';
 import { ActivityIndicator, StyleSheet, View } from 'react-native';
-import type {
-  Hash,
-  LocaleCandidates,
-  Locale,
-  WritableConditionStoreInterface,
-} from 'gt-i18n/internal/types';
-import type { Translation } from 'gt-i18n/types';
+import type { LocaleCandidates } from 'gt-i18n/internal/types';
 import type { NativeConditionStoreParams } from '../condition-store/NativeConditionStore';
-import { NativeConditionStore } from '../condition-store/NativeConditionStore';
-import { getLocale, resolveLocale } from '../utils/getLocale';
+import { NativeGTProvider } from './NativeGTProvider';
 
 export type GTProviderProps = Omit<
   InternalGTProviderProps,
@@ -28,15 +16,6 @@ export type GTProviderProps = Omit<
     loadingFallback?: ReactNode;
   };
 
-type NativeGTProviderProps = Omit<GTProviderProps, 'loadingFallback'>;
-type LocaleTranslations = Record<Hash, Translation>;
-type TranslationSnapshot = Record<Locale, LocaleTranslations>;
-
-const translationPromises = new WeakMap<
-  object,
-  Map<string, Promise<LocaleTranslations>>
->();
-
 export function GTProvider(props: GTProviderProps) {
   const { loadingFallback, ...providerProps } = props;
 
@@ -45,112 +24,6 @@ export function GTProvider(props: GTProviderProps) {
       <NativeGTProvider {...providerProps} />
     </Suspense>
   );
-}
-
-function NativeGTProvider(props: NativeGTProviderProps) {
-  const {
-    locale,
-    region,
-    enableI18n,
-    localeStoreKey,
-    regionStoreKey,
-    enableI18nStoreKey,
-    ...providerProps
-  } = props;
-  const [nativeLocale, setNativeLocale] = useState(() =>
-    resolveLocale(locale ?? getLocale({ localeStoreKey }))
-  );
-  const [conditionVersion, setConditionVersion] = useState(0);
-  const activeLocale = resolveLocale(locale ?? nativeLocale);
-  const localeTranslations = use(loadTranslations(activeLocale));
-  const translations = useMemo<TranslationSnapshot>(
-    () => ({ [activeLocale]: localeTranslations }),
-    [activeLocale, localeTranslations]
-  );
-
-  const i18nStoreRef = useRef<I18nStore | null>(null);
-  if (i18nStoreRef.current == null) {
-    i18nStoreRef.current = new I18nStore();
-  }
-
-  const handleNativeConditionChange = useCallback(
-    (nativeConditionStore: NativeConditionStore) => {
-      if (locale === undefined) {
-        setNativeLocale(nativeConditionStore.getLocale());
-      }
-      setConditionVersion((version) => version + 1);
-    },
-    [locale]
-  );
-
-  const nativeConditionStore = useMemo(
-    () =>
-      new NativeConditionStore({
-        locale: activeLocale,
-        region,
-        enableI18n,
-        localeStoreKey,
-        regionStoreKey,
-        enableI18nStoreKey,
-      }),
-    [
-      activeLocale,
-      region,
-      enableI18n,
-      localeStoreKey,
-      regionStoreKey,
-      enableI18nStoreKey,
-    ]
-  );
-
-  const conditionStore = useMemo<WritableConditionStoreInterface>(
-    () => ({
-      getLocale: nativeConditionStore.getLocale,
-      getRegion: nativeConditionStore.getRegion,
-      getEnableI18n: nativeConditionStore.getEnableI18n,
-      setLocale: (nextLocale: string) => {
-        nativeConditionStore.setLocale(nextLocale);
-        handleNativeConditionChange(nativeConditionStore);
-      },
-      setRegion: (nextRegion: string | undefined) => {
-        nativeConditionStore.setRegion(nextRegion);
-        handleNativeConditionChange(nativeConditionStore);
-      },
-      setEnableI18n: (nextEnableI18n: boolean) => {
-        nativeConditionStore.setEnableI18n(nextEnableI18n);
-        handleNativeConditionChange(nativeConditionStore);
-      },
-    }),
-    [conditionVersion, handleNativeConditionChange, nativeConditionStore]
-  );
-
-  return (
-    <InternalGTProvider
-      {...providerProps}
-      translations={translations}
-      conditionStore={conditionStore}
-      i18nStore={i18nStoreRef.current}
-    />
-  );
-}
-
-function loadTranslations(locale: string): Promise<LocaleTranslations> {
-  const i18nCache = getReactI18nCache();
-  let i18nCacheTranslationPromises = translationPromises.get(i18nCache);
-  if (i18nCacheTranslationPromises == null) {
-    i18nCacheTranslationPromises = new Map();
-    translationPromises.set(i18nCache, i18nCacheTranslationPromises);
-  }
-
-  let promise = i18nCacheTranslationPromises.get(locale);
-  if (promise == null) {
-    promise = i18nCache.loadTranslations(locale).catch((error: unknown) => {
-      i18nCacheTranslationPromises.delete(locale);
-      throw error;
-    });
-    i18nCacheTranslationPromises.set(locale, promise);
-  }
-  return promise;
 }
 
 function DefaultLoadingFallback() {
